+++ .pylintrc
@@ -3,11 +3,7 @@
 #rcfile=
 # Python code to execute, usually for sys.path manipulation such as
 # pygtk.require().
-
-# Address 'maximum recursion' issue
-# https://stackoverflow.com/questions/36496192
-init-hook='import sys; sys.setrecursionlimit(8 * sys.getrecursionlimit())'
-
+#init-hook=
 # Add files or directories to the exclude list. They should be base names, not
 # paths.
 ignore=third_party
