+++ requirements.txt
@@ -1,15 +1,12 @@
-avro == 1.10.0
-avro-python3 == 1.9.2.1
-apache-beam[gcp] == 2.24.0
+apache-beam[gcp] >= 2.22.0
 pandas >= 1.0.4
-# TODO(cezequiel): Support breaking changes in TF Transform 0.25.0
-tensorflow_transform >= 0.22, <= 0.24.1
+tensorflow_transform >= 0.22
 Pillow >= 7.1.2
 coverage >= 5.1
 ipython >= 7.15.0
 nose >= 1.3.7
 numpy < 1.19.0
-pylint == 2.6.0
+pylint >= 2.5.3
 fire >= 0.3.1
 jupyter >= 1.0.0
 tensorflow >= 2.3.1
